Add utility to check for login and admin status

+++ backend/route-callbacks/deleteUser.js
@@ -1,13 +1,21 @@
+import {isLoggedInAdmin} from '../utils';
+
 export function deleteUser(req, res, Users) {
-	const {userId} = req.params;
-	if (userId) {
-		Users.deleteOne({_id: userId}, (error) => {
-			if (error) {
-				return res.json({succes: false, error});
-			}
-			return res.json({success: true});
-		})
+	console.log(req.session);
+	if (isLoggedInAdmin(req)) {
+		const {userId} = req.params;
+		if (userId) {
+			Users.deleteOne({_id: userId}, (error) => {
+				if (error) {
+					return res.json({succes: false, error});
+				}
+				return res.json({success: true});
+			})
+		} else {
+			return res.json({succes: false, error: 'No user id provided'});
+		}
 	} else {
-		return res.json({succes: false, error: 'No user id provided'});
+		return res.json({success: false, error: 'Must be logged in admin to delete user'});
 	}
+
 }
